ChainEndpoint: Use default implementations from CosmosSdkChain (#2807)
* ChainEndpoint: Use default implementations from CosmosSdkChain

Fixes #2806.

This requires changing config() to return a &ChainConfig instead of a
ChainConfig, so the one call site using config() is also updated.

* Fix changelog link


* Undo `init_event_monitor` change
