+++ crates/relayer/src/chain/cosmos.rs
@@ -7,8 +7,7 @@ use core::{
     time::Duration,
 };
 use num_bigint::BigInt;
-use std::cmp::Ordering;
-use std::thread;
+use std::{cmp::Ordering, thread};
 
 use ibc_proto::protobuf::Protobuf;
 use tendermint::block::Height as TmHeight;
@@ -23,15 +22,21 @@ use tonic::{codegen::http::Uri, metadata::AsciiMetadataValue};
 use tracing::{error, instrument, trace, warn};
 
 use ibc_proto::cosmos::staking::v1beta1::Params as StakingParams;
+use ibc_relayer_types::clients::ics07_tendermint::client_state::{
+    AllowUpdate, ClientState as TmClientState,
+};
+use ibc_relayer_types::clients::ics07_tendermint::consensus_state::ConsensusState as TMConsensusState;
 use ibc_relayer_types::clients::ics07_tendermint::header::Header as TmHeader;
 use ibc_relayer_types::core::ics02_client::client_type::ClientType;
 use ibc_relayer_types::core::ics02_client::error::Error as ClientError;
+use ibc_relayer_types::core::ics02_client::events::UpdateClient;
 use ibc_relayer_types::core::ics03_connection::connection::{
     ConnectionEnd, IdentifiedConnectionEnd,
 };
 use ibc_relayer_types::core::ics04_channel::channel::{ChannelEnd, IdentifiedChannelEnd};
 use ibc_relayer_types::core::ics04_channel::packet::Sequence;
 use ibc_relayer_types::core::ics23_commitment::commitment::CommitmentPrefix;
+use ibc_relayer_types::core::ics23_commitment::merkle::MerkleProof;
 use ibc_relayer_types::core::ics24_host::identifier::{
     ChainId, ChannelId, ClientId, ConnectionId, PortId,
 };
@@ -44,19 +49,15 @@ use ibc_relayer_types::core::ics24_host::{
 };
 use ibc_relayer_types::signer::Signer;
 use ibc_relayer_types::Height as ICSHeight;
-use ibc_relayer_types::{
-    clients::ics07_tendermint::client_state::{AllowUpdate, ClientState as TmClientState},
-    core::ics23_commitment::merkle::MerkleProof,
-};
-use ibc_relayer_types::{
-    clients::ics07_tendermint::consensus_state::ConsensusState as TMConsensusState,
-    core::ics02_client::events::UpdateClient,
-};
 
-use crate::account::Balance;
 use crate::chain::client::ClientSettings;
+use crate::chain::cosmos::batch::sequential_send_batched_messages_and_wait_commit;
+use crate::chain::cosmos::batch::{
+    send_batched_messages_and_wait_check_tx, send_batched_messages_and_wait_commit,
+};
 use crate::chain::cosmos::encode::key_entry_to_signer;
 use crate::chain::cosmos::fee::maybe_register_counterparty_payee;
+use crate::chain::cosmos::gas::{calculate_fee, mul_ceil};
 use crate::chain::cosmos::query::account::get_or_fetch_account;
 use crate::chain::cosmos::query::balance::{query_all_balances, query_balance};
 use crate::chain::cosmos::query::denom_trace::query_denom_trace;
@@ -70,10 +71,6 @@ use crate::chain::cosmos::types::config::TxConfig;
 use crate::chain::cosmos::types::gas::{
     default_gas_from_config, gas_multiplier_from_config, max_gas_from_config,
 };
-use crate::chain::cosmos::{
-    batch::sequential_send_batched_messages_and_wait_commit,
-    gas::{calculate_fee, mul_ceil},
-};
 use crate::chain::endpoint::{ChainEndpoint, ChainStatus, HealthCheck};
 use crate::chain::requests::{Qualified, QueryPacketEventDataRequest};
 use crate::chain::tracking::TrackedMsgs;
@@ -82,19 +79,15 @@ use crate::config::ChainConfig;
 use crate::consensus_state::{AnyConsensusState, AnyConsensusStateWithHeight};
 use crate::denom::DenomTrace;
 use crate::error::Error;
-use crate::event::monitor::{EventMonitor, EventReceiver, TxMonitorCmd};
+use crate::event::monitor::{EventReceiver, TxMonitorCmd};
 use crate::event::IbcEventWithHeight;
 use crate::keyring::{KeyEntry, KeyRing};
 use crate::light_client::tendermint::LightClient as TmLightClient;
 use crate::light_client::{LightClient, Verified};
 use crate::misbehaviour::MisbehaviourEvidence;
 use crate::util::pretty::{PrettyConsensusStateWithHeight, PrettyIdentifiedChannel};
-use crate::{
-    chain::cosmos::batch::{
-        send_batched_messages_and_wait_check_tx, send_batched_messages_and_wait_commit,
-    },
-    util::pretty::{PrettyIdentifiedClientState, PrettyIdentifiedConnection},
-};
+use crate::util::pretty::{PrettyIdentifiedClientState, PrettyIdentifiedConnection};
+use crate::{account::Balance, event::monitor::EventMonitor};
 
 use super::requests::{
     IncludeProof, QueryChannelClientStateRequest, QueryChannelRequest, QueryChannelsRequest,
@@ -682,10 +675,6 @@ impl ChainEndpoint for CosmosSdkChain {
         Ok(())
     }
 
-    fn id(&self) -> &ChainId {
-        &self.config().id
-    }
-
     fn keybase(&self) -> &KeyRing {
         &self.keybase
     }
@@ -788,29 +777,8 @@ impl ChainEndpoint for CosmosSdkChain {
     }
 
     /// Get the chain configuration
-    fn config(&self) -> ChainConfig {
-        self.config.clone()
-    }
-
-    /// Get the signing key
-    fn get_key(&mut self) -> Result<KeyEntry, Error> {
-        crate::time!("get_key");
-
-        // Get the key from key seed file
-        let key = self
-            .keybase()
-            .get_key(&self.config.key_name)
-            .map_err(|e| Error::key_not_found(self.config.key_name.clone(), e))?;
-
-        Ok(key)
-    }
-
-    fn add_key(&mut self, key_name: &str, key: KeyEntry) -> Result<(), Error> {
-        self.keybase_mut()
-            .add_key(key_name, key)
-            .map_err(Error::key_base)?;
-
-        Ok(())
+    fn config(&self) -> &ChainConfig {
+        &self.config
     }
 
     fn ibc_version(&self) -> Result<Option<semver::Version>, Error> {
